python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy>=1.26.0
google-generativeai==0.3.0
python-docx==1.1.0
PyPDF2==3.0.1
python-multipart==0.0.6

